AposSchema v-model and remove ApostropheSchemaEditor
